requirements.txt
streamlit
python-dotenv
phi-client
openai 